create asome api

+++ src/app/modules/collection/other-collection-Name/otherCollectionName.services.ts
@@ -11,8 +11,12 @@ const getallcollectionName = async () => {
   return await prisma.otherCollectionName.findMany({
     orderBy: { createdAt: "desc" },
   });
-}
+};
 const deletecollectionName = async (id: string) => {
+  const collectionName = await prisma.otherCollectionName.findUnique({
+    where: { id },
+  });
+  if (!collectionName) throw new Error("Collection Name id not found");
   return await prisma.otherCollectionName.delete({
     where: { id },
   });
@@ -21,11 +25,15 @@ const updatecollectionName = async (
   id: string,
   data: Partial<IOtherCollectionName>
 ) => {
+  const isExits = await prisma.otherCollectionName.findUnique({
+    where: { id },
+  });
+  if (!isExits) throw new Error("Collection Name id not found");
   return await prisma.otherCollectionName.update({
     where: { id },
     data: data,
   });
-};  
+};
 
 export const OtherCollectionNameServices = {
   createcollectionName,
